fix(components): [date-picker] circular dependency (#8093)
